backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
google-genai==1.0.0
python-multipart==0.0.12
aiosqlite==0.20.0
python-dotenv==1.0.0

# Development tools
black==24.10.0
ruff==0.7.4
